refactor v2 trace sort helpers
